Add files via upload
This code is the building blocks that were used for the final file of code, some tweaks remain on each file but the basic function is there. Due to time constraints each graph has its own lines, did not have time to think of a way to input the data into a function and produce a graph for each one.